Update hpc.rst
When performing a ``conda update conda`` it may create a ``(base)`` environment depending on the user choice during install, so adding a note about that and how to prevent this behavior at initialization, if desired.  As of conda 4.4, ``conda activate`` is the preferred, way to activate an environment, supposed to be faster and work across different operating systems.