Minor fixes on hpc setup guide for cheyenne (#592)
* Update hpc.rst

One has to create the ``jupyter_notebook_config.json``  before calling ``jupyter notebook password`` otherwise you get a file not found error.

* Update hpc.rst

When performing a ``conda update conda`` it may create a ``(base)`` environment depending on the user choice during install, so adding a note about that and how to prevent this behavior at initialization, if desired.  As of conda 4.4, ``conda activate`` is the preferred, way to activate an environment, supposed to be faster and work across different operating systems.

* Update hpc.rst

rephrasing last change in text.

* Fixing distributed.yaml edit

.config/dask/distributed.yaml does not exist before ``dask.distributed`` or
``dask-jobqueue`` is first imported. Moreover it may be important to point out that it is
 a commented out version. I added what I consider the simplest way to do this first import 
and updated to the current "look" of the distributed.yaml